==> entity_linking/03_add_dbpedia_links.ipynb <==
# Add DBpedia Links

Add DBpedia links to software in the table of additional information. 
Because we qeury DBpedia automatically this information does not need to be gathered by hand but can be inserted automatically which we do here.

import pandas as pd
dbpedia_names = pd.read_csv('dbpedia_software_long.csv.gz', compression='gzip')
info_table = pd.read_csv('../data/software_reasoning_list_total_neu.csv')

new_column = []
for s_name in info_table['linked_name']:
    rows = dbpedia_names.loc[dbpedia_names['unique']==s_name]
    rows = rows[['unique', 'uri']]
    rows = rows.drop_duplicates()
    if len(rows) > 1:
        print("Too many entries for {}".format(s_name))
        new_column.append('')
    elif len(rows) < 1:
        new_column.append('')
    else:
        new_column.append(rows.iloc[0]['uri'])

info_table['DBpedia'] = new_column
cols = info_table.columns.tolist()
cols = cols[0:2] + cols[-1:] + cols[2:-1]
info_table_2 = info_table[cols]

info_table_2.to_csv("software_reasoning_final_production_model.csv.gz", compression="gzip")
